fix(ci): stabilize eval-mirror branch gates
