+++ crates/tsz-checker/src/state/type_environment/lazy_flow_mirror.rs
@@ -1,0 +1,42 @@
+//! Lazy-resolution writes that mirror evaluator env updates into the flow env.
+
+use crate::state::CheckerState;
+use tsz_solver::TypeId;
+use tsz_solver::def::DefId;
+
+impl CheckerState<'_> {
+    /// Insert `type_id` for `def_id` into both type environments, carrying type
+    /// params when present. Safe to call during recursive resolution; failed
+    /// borrows are logged.
+    pub(super) fn try_insert_def_in_type_env(&mut self, def_id: DefId, type_id: TypeId) {
+        let params = self.ctx.get_def_type_params(def_id).unwrap_or_default();
+        match self.ctx.type_env.try_borrow_mut() {
+            Ok(mut env) => env.insert_def_with_params(def_id, type_id, params.clone()),
+            Err(e) => tracing::warn!(
+                target_env = "type_env",
+                error = ?e,
+                "try_insert_def_in_type_env: borrow failed; insert skipped"
+            ),
+        }
+        match self.ctx.type_environment.try_borrow_mut() {
+            Ok(mut env) => env.insert_def_with_params(def_id, type_id, params),
+            Err(e) => tracing::warn!(
+                target_env = "type_environment",
+                error = ?e,
+                "try_insert_def_in_type_env: borrow failed; insert skipped"
+            ),
+        }
+    }
+
+    pub(super) fn mirror_application_def_resolution(
+        &self,
+        def_id: Option<DefId>,
+        resolved: TypeId,
+        type_params: &[tsz_solver::TypeParamInfo],
+    ) {
+        if let Some(def_id) = def_id {
+            self.ctx
+                .mirror_def_in_type_environment(def_id, resolved, type_params);
+        }
+    }
+}
